a11y: announce shapes as they're visited (#5773)
Building off of https://github.com/tldraw/tldraw/pull/5634 and
https://github.com/tldraw/tldraw/pull/5761 this is adding a11y live text
to be read aloud when visiting a shape.

We add an overridable method for shapes to customize this called
`getAriaLiveText`.
Furthermore, we lay the groundwork here to start letting media shapes
have `altText`.
Drive-by fix of `heart` being missing in `geo-styles` list.
Also, drive-by fix of us calling our Image button "Asset" (what are we
selling financial instruments here? :P) "Media" is a better word for
this button, more human.

Some of the i18n translation is funky. It's a shortcoming of our current
system that we don't support interpolation :-/
It sucks, and we'll revisit in the future.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [x] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- a11y: announce shapes as they're visited

---------

Co-authored-by: alex <alex@dytry.ch>