Fix e2e tests. (#5844)
This implements `run` api that zero has added.

Also had to make some of our apis async. Example issue that this fixes
is duplicating files. There we first called `createFile` and then
navigated to it. But since `insertWithFileState` is `async` we got the
file state update from the navigation before the file was actually
created, which caused some issues.

### Change type

- [x] `bugfix`