Fix flyio app prunning and deployment. (#5834)
Looks like we need to pass the `-y` flag to [accept the
confirmation](https://fly.io/docs/flyctl/apps-destroy/#options) when
pruning the apps.
[Failure](https://github.com/tldraw/tldraw/actions/runs/14297908654/job/40067462974#step:6:346)
of the last run.

Preview deploys were also failing to start correctly because of
permissions deployment issue. We tried to deploy permissions before
deploying zero, which failed since the zero hasn't had the chance to
create the tables yet.

### Change type

- [x] `bugfix`