Deploy zero on demand. (#5885)
We don't want to deploy zero backends for PRs that have nothing to do
with the dotcom app. This should help with that by:
* Not deploying zero by default.
* Deploying zero to flyio or sst if we have
`preview-flyio-zero-deploy-please` or `preview-sst-zero-deploy-please`
labels on the pr.
* It checks for changes to `apps/dotcom` folder. If there are any it
automatically adds the `preview-flyio-zero-deploy-please` label. So we
get automatic deploys when it's highly likely that we'd want them.

Also moved the zero deployment right after migrating the db and before
we deploy sync worker.

### Change type

- [x] `improvement`