Add overlays component (#5952)
Tldraw has two main places to add stuff to the canvas: the `OnTheCanvas`
and the `InFrontOfTheCanvas` components.

`OnTheCanvas` follows the camera, and allows you to place content within
shapes and below overlays. `InFrontOfTheCanvas` does not follow the
camera, and allows you to place content above overlays.

This diff introduces a 3rd slot, `Overlays`, which exists between the
two. It follows the camera, is above the shapes (and overlays, although
that's configurable) but is below the `InFrontOfTheCanvas` layer.

I've felt the need for something like this a few times, but the main
motivation is from the hints we show with elbow arrows. For these, the
existing things aren't really appropriate. Selection foreground (for
example) can't be used because it's those overlays show when the
selection is hidden, and are on a different shape to the one selected
(the arrow's target).

### Change type

- [x] `api`

### Release notes

- A new `Overlays` component lets you add overlays on top of the canvas
that still follow the camera.