Indicators perf (#5821)
This PR improves the rendering for indicators. Now that we're leaving
them in the DOM (and hiding them) to avoid expensive mounting /
unmounting, I've seen the rendering cost of indicators effecting
performance. This really only happens when dragging large numbers of
shapes around, though on slower computers it might show up with smaller
numbers of shapes.

In this PR, we do two things:
- we now don't update the position of indicators if they're hidden
- we only render the indicator if its associated shape's props have
changed

And a third, secret thing:

- for geo shapes, we render the "solid" style indicator when zoomed out

### Change type

- [x] `improvement`

### Test plan

1. Make many shapes
2. Drag them about the canvas

### Release notes

- Improved performance when dragging many shapes at once

---------

Co-authored-by: David Sheldrick <d.j.sheldrick@gmail.com>