Fix clickable area on form rows (#6007)
This PR fixes a bug where the entire row in the exports menu would be
clickable, though its cursor was not set to `pointer`. Actually the row
should not have been clickable; and if it is, then we need to indicate
it and disambiguate the hover state from the hover of the tooltip.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Hover items in the export menu
2. Hover / click the label of the export menu

### Release notes

- Fixed a bug with the clickable area of items in the export menu