[VSCode] Allow users to drag files from the sidebar. (#5956)
This allows users to drag image files from the sidebar into the editor.
In this case we don't get the files (so the files content handler does
not get called). Instead we get an url prefixed by `file://`. We have to
send the url to the extension, since the webview runs in a sandbox and
does not have file system access. The extension gets the file and sends
it back.

Resolves ENG-3111

### Change type

- [x] `bugfix`