Improve performance of a programatically rotating shape when it is selected (#6023)
Looks like the main slowdown is downloading the updated cursors over and
over again and then rendering them. Check how many requests the browser
makes in the before case.

### Before


https://github.com/user-attachments/assets/34a6f317-e605-4367-b8b7-f6e26cef07e6

### After


https://github.com/user-attachments/assets/2ab46925-3f66-446a-b4d2-c4715b3d2fe6


Hovering still results in slowdown, but that's just the the result of
updating the cursor that many times.


https://github.com/user-attachments/assets/5ea221e5-3bbc-4da0-9105-c5507af77ee4


### Change type

- [x] `improvement`

### Release notes

- Improve the performance of shape rotation. This slowdown could be seen
when the shape was rotated programmatically and you selected it. Or in a
multiplayer scenario if you selected a shape that somebody else was
rotating.