arrows: create min-width for label indicator (#6029)
With elbow arrows, we have a lot more vertical segments. When starting
editing on the label, you can't see the cursor blinking b/c it overlaps
with the shape. This makes the rectangle indicator show up when editing,
even if we don't have text in the label yet.

I would view the diff with "hide whitespace" btw.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- arrows: create min-width for label indicator

---------

Co-authored-by: alex <alex@dytry.ch>