Prevent announcements during interactions (#6041)
This PR fixes a bug causing accessibility announcements to occur too
often. The code as written was firing during interactions (translating a
shape, etc) and on hover/unhovers, due to the announcement function
being captured. This narrows it to only when appropriate.

This announcement isn't free so if there's a way to know whether the
user is consuming this announcements, we can turn them off until that's
true.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Turn on screenreader
2. Drag a shape around
3. Select it
4. Deselect it

### Release notes

- Fixed a bug causing too many screenreader announcements