Snap pasted text and embed content to the grid (#6020)
When creating a shape from pasted text or embed content, it wasn't
snapped to the grid. Now it is if the grid is enabled.

Describe what your pull request does. If you can, add GIFs or images
showing the before and after of your change.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Enable the grid.
2. Put some text or a URL to embeddable content on the clipboard.
3. Paste it on the canvas.
4. Observe that it's now snapped to the grid.

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- When pasting text or embeddable links, snap the created shape to the
grid if the grid is enabled.