Make sure the file exists before uploading it to the bucket. (#6027)
This makes sure that the file exists before we upload an asset to the R2
bucket. Otherwise you could upload assets and not have them be
associated with a file.

This also uses a queue for inserting the asset <> file association.
Otherwise the insert might fail and we might still not correctly
associate assets to files. Using a queue allows us to retry the inserts
with an exponential backoff.

### Change type

- [x] `improvement`