a11y: fix popover focus (#5994)
(this is on top of another PR, the Cmd+Enter one so it's more easily
visible how the change affects tabbing through shapes)

This selects the first, visible, non-disabled button in a popover when
opened. I made the default true but I'm not opposed and could be
convinced that the default should be false.

- Drive-by fix here to add a placeholder to the Edit link dialog.
- also, a fix for the Slider, in MobileStylePanel stealing focus when
being opened.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- a11y: focus first item in popover when opened