analytics: lay foundation to unify across tldraw properties (#6089)
tldraw is a business and we're looking to get a consistent view of our
user's journeys across the various websites/products/experiments we
host.

This includes:
- the main website ("dotcom")
- the docs (.dev)
- computer.tldraw.com
- teach.tldraw.com
- makereal.tldraw.com
- examples
- etc.

We have a monorepo for .com/.dev/examples, but our other work is in
other repos. The tricky thing there becomes:
- being able to consisently ask for consent.
- and have consistent tracking metrics/service without having to open up
4 PRs in separate repos every time we update our analytics code.

Further requirements:
- On top of the existing analytics we have, which is Vercel + PostHog
(in places), our marketing team would like to also add GA4 on the
frontend.
- We were initially going to use Twilio's Segment as a relay. While that
worked well for most things (worked perfectly with Poshog, ahem), GA4
did not play nicely.
- Given that and the fact that we'll be sunsetting Vercel analytics,
we'll just run GA4 and PostHog side-by-side.
- I'm also in discussion to maybe see if PostHog can just be used
instead of GA4 since it seems to be much better at what it does than
GA4.

So, this PR introduces building a JS script that we compile, using Radix
+ PostHog + GA4 to:
- have a consistent Opt out/Accept consent experience
- have a Privacy Dialog that lets you update that choice, as required by
us
- and include the appropriate analytics (GA4, PostHog) when consent is
given

Still more work to be done here but this is the initial groundwork. 

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [x] `other`