Consolidate mutation logic (#6075)
# Consolidate Mutation Logic

This PR makes it so that we don't have separate mutation and
authorization logic for zero. Making this possible involved cleaning up
and simplifying a lot of other parts of the codebase. I'm gonna walk
though the changes here:

1. **Simplifying the ZStoreData type**


https://github.com/tldraw/tldraw/blob/c0d0f8a755fdff0e9f2f33617dc45109d9e8f400/packages/dotcom-shared/src/types.ts#L78-L90
  
Previously this type used camelCase versions of our table names, and
also for the 'user' table it used a single nullable row rather than an
array of rows. This required a lot of special casing in the code. I
changed this to use snake_case for the table names, and made the user
table an array of rows. This means that we can now use the same logic
for all tables, and we don't need as much glue code.

Making this change naively would break older clients, so I added a
down-migration step when sending the 'initial_data' message.


https://github.com/tldraw/tldraw/blob/c0d0f8a755fdff0e9f2f33617dc45109d9e8f400/apps/dotcom/sync-worker/src/TLUserDurableObject.ts#L191-L201

2. **Simplifying the client CRUD/Query logic, making them disposable**

As mentioned above, having a simpler ZStoreData type means that we can
use the same logic for all tables so I deleted a lot of the boilerplate
code while moving the Client CRUD/Query logic into their own classes
[ClientCRUD](https://github.com/tldraw/tldraw/blob/c0d0f8a755fdff0e9f2f33617dc45109d9e8f400/apps/dotcom/client/src/tla/app/ClientCRUD.ts#L5)
and
[ClientQuery](https://github.com/tldraw/tldraw/blob/c0d0f8a755fdff0e9f2f33617dc45109d9e8f400/apps/dotcom/client/src/tla/app/ClientQuery.ts#L4).

These classes get instantiated per-mutation, and receive an abort signal
so we can throw errors if they are used after the mutation has
'finished' and the crud tracking has stopped.

3. **Replacing _doMutate ad-hoc CRUD logic with a ZQL-style CRUD class
backed by postgres**

To eventually use zero-style mutators in our custom backend we need to
implement a CRUD class that points to postgres.
   

https://github.com/tldraw/tldraw/blob/a1ea48518d89627f522b06a31d6f8d5cc0986ccc/apps/dotcom/sync-worker/src/zero/ServerCrud.ts#L9

We can also use this to handle legacy 'update'-style mutations from
older clients during a transition period.


https://github.com/tldraw/tldraw/blob/d043ca6e8bd2353f59ea4dac1980bebd42f66807/apps/dotcom/sync-worker/src/TLUserDurableObject.ts#L362-L366

This meant moving one bit of ad-hoc validation logic that was in
_doMutate into the `assertValidMutation` method. Namely, checking that
MAX_FILES hasn't been exceeded. And `assertValidMutation` itself was
moved to a separate file and marked as legacy because it's only needed
for the legacy update-style mutations.

4. **Sending 'mutator'-style mutations to the server**

In order to let us eventually consolidate our authorization logic, we
need to be sending 'mutator'-style mutations to the server instead of
'update'-style mutations. Then our custom mutators will be entirely
responsible for auth checks.


https://github.com/tldraw/tldraw/blob/d043ca6e8bd2353f59ea4dac1980bebd42f66807/packages/dotcom-shared/src/types.ts#L157-L168

In order to handle the mutators on the server, we need to implement an
polyfill of the zql query interface on the server backed by postgres.


https://github.com/tldraw/tldraw/blob/d043ca6e8bd2353f59ea4dac1980bebd42f66807/apps/dotcom/sync-worker/src/zero/ServerQuery.ts#L5

## Testing

I've manually tested the downgrading by running an older version of the
client against this version of the backend.

e2e tests are all passing.

### Change type

- [x] `other`