Use PathBuilder for geo & line shapes (#6112)
Before #5572, we had two geometric style shapes that required inky
rendering: lines and geo shapes. Both of these shapes implemented their
own inky algorithm, plus approaches for turning into dashed and dotted
lines too. For geo shapes, this was particularly bad - many geo shapes
needed their own completely custom implementations for inky & dashed
lines, as well as for geometry.

In #5572, we introduced a third: elbow arrows (straight and arc arrows
don't get draw-style rendering currently). To facilitate this, that PR
introduced the foundations of a new abstraction for working with paths
in tldraw: `PathBuilder`. `PathBuilder` has an API similar to that of
the webs
[`Path2D`](https://developer.mozilla.org/en-US/docs/Web/API/Path2D). It
allowed the construction of a path using familiar, canvas-like commands,
that could then be converted into `Geometry2d` or SVG - including in
inky and dashed styles. The plan when I introduced that API was always
to use it for line and geo shapes too, but that was out of scope of the
initial elbow arrows PR.

This PR revamps `PathBuilder` to suit the needs of lines and geo shapes,
and replaces all the many custom implementations of draw and dash style
rendering with the single approach defined in `PathBuilder`. Among other
things, this means it's now much easier to add new geo shapes (just add
the definition in `getGeoShapePath` and the geometry / rendering will be
handled automatically) & much easier for developers to create shapes
that use tldraws draw/dash styles (define the path using `PathBuilder`,
then call `toSvg` with the appropriate methods).

There are some extremely subtle differences in rendering as a result of
this PR as there were subtle differences between many of the previous
algorithms.

### Change type

- [x] `api`


### Release notes

- Introduce `PathBuilder`, a new API to make constructing tldraw-style
inky and dashed SVGs much easier.

---------

Co-authored-by: huppy-bot[bot] <128400622+huppy-bot[bot]@users.noreply.github.com>