Improve resizing frames (#6128)
This PR improves the logic around resizing frames, so that the children
of frames do not resize when a frame is being resized with other shapes

![Kapture 2025-05-26 at 07 51
36](https://github.com/user-attachments/assets/34b14967-3c88-417c-b1c5-ad0470fa0045)

This PR also improves how accelKey is handled when resizing frames /
multiple frames:

![Kapture 2025-05-26 at 08 05
55](https://github.com/user-attachments/assets/0f5f7e17-5408-4b4f-9a67-50a56ed3e846)


In this PR, we:
- clean up the resizing snapshot collection code
- add a `canResizeChildren` helper to ShapeUtil
- use `canResize` and `canResizeChildren` in the new resizing snapshot
collection code
- handle accelKey resizing with frames when multiple frames are selected

## Background 

Resizing a frame does not resize its children. This is true and good.
However, previously, when a frame was in a selection that included other
shapes, resizing the frame would also resize its children. It would not
resize _all_ of its descendants, just the direct children.

I did explore an alternative where all of the descendants would resize,
however that felt even less appropriate.

### Change type

- [x] `bugfix`
- [x] `improvement`

### Release notes

- Fixed several bugs with resizing frames when more than one frame is
selected.