[hotfix] don't show deleted files in sidebar (#6144)
Somehow we have not been effectively cleaning up people's own file
states when they delete their own files. This PR does not fix that but
it does prevent the downstream issue where we still show deleted files
in the sidebar.

In main this is no longer a problem, due to a recent refactoring, but we
can't cherry-pick that refactoring so it's either hotfix or do a deploy.
This is the hotfix.

### Change type

- [x] `other`