Fix pruning of queues and workers (#6143)
Fixes pruning of workers and queues:
* The main problem was that cloudflare api requires you to specify queue
id and not the name.
* For the queue consumer you have to specify the consumer id - the
worker name does not work on this case. We now fetch the info for all
the queues. We then use this info to get queue ids and queue consumer
ids.
* Also fixed the issue for cases where we deleted the main sync worker
before the image optimizier. That didn't work since the image optimizer
had a service binding to the main worker.

Alternative would be to use the `force=true` option for the worker
delete request, which should delete all the associated services. It
looks like that doesn't work with queues though 🤷‍♂️

### Change type

- [x] `bugfix`