Configurable VideoShapeUtil autoplay (#6070)
This pull request is a second attempt to resolve
https://github.com/tldraw/tldraw/issues/5716.

It was mentioned during a previous attempt
(https://github.com/tldraw/tldraw/pull/6016) by @mimecuvalo that the
preferred approach to defining behaviour like this was to allow for
configuration of the shape util in question rather than a global
setting. This PR allows for `VideoShapeUtil` to be configured using the
`defaultAutoplay` boolean property.

Some things to note:

- This parameter sets the `playing` prop based on the value of
`defaultAutoplay` and then uses that prop in the rendering of the shape.
This choice was made because a) that prop is currently just hard coded
to `true` and not actually used anywhere and b) Adding an `autoPlay`
prop would introduce additional complexity in additional code
requirements such as migrations.
- I hope this is sufficient to allow this behaviour to be introduced
because as noted in the closed PR - the current situation of multiple
videos all autoplaying at once is pretty egregious, so please use this
as a starting point for further improvements that may allow this PR to
go in.

Thanks in advance.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Note all videos currently added to a canvas autoplay
2. Configure your canvas to use a configured `VideoShapeUtil` e.g.
`VideoShapeUtil.configure({ defaultAutoplay: true }`
3. Note that videos now do not autoplay.

### Release notes

- Allows for `VideoShapeUtil` autoplay behaviour to be configured using
the `defaultAutoplay` boolean property.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>