Tiny improve Box (#6175)
The `Box` class is part of a lot of hot routes. When I wrote the class,
I used a lot of getters. For example, `box.maxX` was a getter that would
return `box.x + box.w`. In very hot routes, like snapping, the overhead
of these getters was enough that accessing the properties directly would
be faster. These getters are still here but in this PR they're not used
internally.

We could further speed things up with caching (as we do with geometry
classes) though in many cases the box itself is being recreated on every
frame anyway. A good counter example would be snapping, where we create
new Vecs on every frame for every snappable shape's center and corners,
which is slow.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

- [x] Unit tests

### Release notes

- Tiny improvements to the `Box` class.