Only re-evaluate fonts when props.richText changes (#6165)
As feared, the font recognition and caching code that we wrote for rich
text has some ugly performance effects, especially when resizing shapes.
This is because any change to the props of a shape with a label (text,
note, geo) would cause the font cache to bust and for us to do the work
of scanning the text content for new fonts.

![Kapture 2025-06-01 at 23 49
121](https://github.com/user-attachments/assets/28330e55-b7da-40bf-835e-0a2a95577800)

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create a few hundred shapes
2. Resize them

### Release notes

- Improved performance when resizing or editing many shapes.