Remove `getShapePageGeometry` (#6202)
This PR removes `getShapePageGeometry`. 

### Background

Geometries are complex and so implement lots of internal caching to
avoid recomputing vertices or bounds. These properties are often
referenced many times during interactions like snapping, selecting, or
translating.

Geometry transforms were added in
https://github.com/tldraw/tldraw/pull/5754 while working on elbow
arrows. When a geometry is transformed, its vertices etc and other
things are no longer cached and they're recomputed instead. This is fine
if the transform is part of the creation of the shape (as in the
PathBuilder2d class) but not if we're creating new geometries as we
access them, which is done by `getShapePageGeometry`.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`