Cache rich text -> html (#6252)
This PR improves performance for resizing shapes with text in them and
other situations where text is recomputed. Part of this is measurement,
which is annoying slow, but the other part is just generating the raw
html text, which is also relatively slow and which this PR improves
through a weak cache.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Create lots of shapes with labels
2. Resize them

### Release notes

- Improved performance of resizing