Allow downloading of the original image / video (#6221)
Allow users to download original assets. Didn't make this generic (like
adding `canDownloadOriginal` to shape util), but if the users would want
it we could support it down the line.

### Change type

- [x] `feature`

### Release notes

Allow users to download the original asset via a built in action.
Supports `TLImageShape` and `TLVideoShape`. Users can access this from
the main menu as well as the context menu when only shapes that support
downloading are selected.