a11y: allow switching back-and-forth from the style panel (#6286)
When using the Cmd+Enter shortcut, you can then adjust styles in the
style panel.
However, the issue is that it's very tricky to get back to the shape you
were manipulating. Hitting Escape takes you back to an unfocused canvas
at the moment.
This change makes it so that you can hit Escape and focus the container
again properly.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- a11y: allow switching back to the shape when doing adjusting styles.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>