Support configuring when arrows should be exact and/or bind (#6290)
By default arrows stop at the edge of the target shape and bind to it.
To stop the arrow at the pointer location you have to hold down alt and
to avoid binding you have to hold down ctrl. These bindings were hard
coded.

In our application we'd like the default to be that it stops at the
pointer, and to stop at the edge of the shape you would hold down alt.
To support this I added two new options to the arrow util. To make it
more generic, I made these options be functions that let's you control
these behaviors as you like.

### Change type

- [ ] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [x] `api`
- [ ] `other`

### Test plan

1. Configure the arrow util, e.g.
```
ArrowShapeUtil.configure({
	shouldBeExact: (editor) => !editor.inputs.altKey,
})
```
2. Create any shape
3. Create an arrow from the shape
4. Observe that it starts/stops at the pointer location by default, and
at the edge of the shape when you hold down alt.

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Support configuring when arrows should be exact and/or bind

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>