Fix LoadingScreen context usage (#6309)
The LoadingScreen component was recently altered to include the tldraw
ui context, so that the spinner could have a localized aria label, only
problem is that the ui context provider was not configured according to
the user's tldraw config so it was ignoring asset url overrides for the
translation files, as well as triggering loading of the 0_merged.svg
from the default cdn url.

### Change type

- [x] `bugfix`

### Release notes

- Fixed a bug where tldraw was ignoring custom asset URL configuration
for the LoadingScreen component, causing spurious requests to
cdn.tldraw.com.