Only bail to mark if id is set
The bailToMark function is sometimes called with an empty string. An
example of that is from the text tool pointing state when max shapes is
reached. Then it cancels and calls bailToMark with an empty string
because markId hasn't been set yet.

When this happens, it would undo to the previous state in the history.
This could for instance remove the shape you just created. So to prevent
this, make sure only to undo if the id is set to a non-empty value.