Prevent crash when creating shapes when max shapes is reached (#6315)
The non-null assertion after creating the note shape caused it to crash
when max shapes is reached, as the shape would not exist after
createShape in that case.

I looked at how the text shape handles this, and handled it similarly.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

1. Set `maxShapesPerPage: 0` in `CustomOptionsExample.tsx`
2. Open the example and try to create a note.

Before this PR it would crash, now it shows the max shapes reach toast.

- [ ] Unit tests
- [ ] End to end tests

### Release notes

- Prevent crash when creating notes when max shapes is reached

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>